book-recommender-phase1-backend/requirements.txt
flask
faiss-cpu
sentence-transformers
requests